requirements.txt
fastapi
uvicorn
python-multipart
pydantic
dotenv
langgraph
google-genai
openai
celery
redis
psycopg2-binary
httpx